+++ internal/common/interfaces/interfaces.go
@@ -1,7 +1,5 @@
 package interfaces
 
-import "golang.org/x/exp/constraints"
-
 // DeepCopier represents object that can be deep-copied.
 type DeepCopier[T any] interface {
 	// DeepCopy returns a deep copy of the object.
@@ -12,9 +10,4 @@ type DeepCopier[T any] interface {
 type Equaler[T any] interface {
 	// Returns true if both objects are equal.
 	Equal(T) bool
-}
-
-// Number represents any integer or floating-point number.
-type Number interface {
-	constraints.Integer | constraints.Float
 }
