+++ internal/scheduler/configuration/configuration.go
@@ -1,6 +1,8 @@
 package configuration
 
 import (
+	"maps"
+	"slices"
 	"time"
 
 	v1 "k8s.io/api/core/v1"
@@ -10,7 +12,6 @@ import (
 	commonconfig "github.com/armadaproject/armada/internal/common/config"
 	grpcconfig "github.com/armadaproject/armada/internal/common/grpc/configuration"
 	profilingconfig "github.com/armadaproject/armada/internal/common/profiling/configuration"
-	"github.com/armadaproject/armada/internal/common/types"
 	"github.com/armadaproject/armada/internal/server/configuration"
 	"github.com/armadaproject/armada/pkg/client"
 )
@@ -139,7 +140,7 @@ type SchedulingConfig struct {
 	// Map from priority class names to priority classes.
 	// Must be consistent with Kubernetes priority classes.
 	// I.e., priority classes defined here must be defined in all executor clusters and should map to the same priority.
-	PriorityClasses map[string]types.PriorityClass `validate:"dive"`
+	PriorityClasses map[string]PriorityClass `validate:"dive"`
 	// Jobs with no priority class are assigned this priority class when ingested by the scheduler.
 	// Must be a key in the PriorityClasses map above.
 	DefaultPriorityClassName string
@@ -235,6 +236,55 @@ type SchedulingConfig struct {
 	DefaultPoolSchedulePriority int
 }
 
+type AwayNodeType struct {
+	// Priority is the priority class priority that the scheduler should use
+	// when scheduling "away" jobs of this priority class on the the node type
+	// referenced by WellKnownNodeTypeName; it overrides the Priority field of
+	// PriorityClass.
+	Priority int32 `validate:"gte=0"`
+	// WellKnownNodeTypeName is the Name of the WellKnownNodeType in question.
+	WellKnownNodeTypeName string `validate:"required"`
+}
+
+type PriorityClass struct {
+	Priority int32 `validate:"gte=0"`
+	// If true, Armada may preempt jobs of this class to improve fairness.
+	Preemptible bool
+	// Limits resources assigned to jobs of this priority class.
+	// Specifically, jobs of this priority class are only scheduled if doing so does not exceed this limit.
+	MaximumResourceFractionPerQueue map[string]float64
+	// Per-pool override of MaximumResourceFractionPerQueue.
+	// If missing for a particular pool, MaximumResourceFractionPerQueue is used instead for that pool.
+	MaximumResourceFractionPerQueueByPool map[string]map[string]float64
+	// AwayNodeTypes is the set of node types that jobs of this priority class
+	// can be scheduled on as "away" jobs (i.e., with reduced priority).
+	//
+	// The scheduler first tries to schedule jobs of this priority class as
+	// "home" jobs, and then tries the elements of this slice in order.
+	AwayNodeTypes []AwayNodeType `validate:"dive"`
+}
+
+func (priorityClass PriorityClass) Equal(other PriorityClass) bool {
+	if priorityClass.Priority != other.Priority {
+		return false
+	}
+	if priorityClass.Preemptible != other.Preemptible {
+		return false
+	}
+	if !maps.Equal(priorityClass.MaximumResourceFractionPerQueue, other.MaximumResourceFractionPerQueue) {
+		return false
+	}
+	if len(priorityClass.MaximumResourceFractionPerQueueByPool) != len(other.MaximumResourceFractionPerQueueByPool) {
+		return false
+	}
+	for k, v := range priorityClass.MaximumResourceFractionPerQueueByPool {
+		if !maps.Equal(v, other.MaximumResourceFractionPerQueueByPool[k]) {
+			return false
+		}
+	}
+	return true
+}
+
 const (
 	DuplicateWellKnownNodeTypeErrorMessage     = "duplicate well-known node type name"
 	AwayNodeTypesWithoutPreemptionErrorMessage = "priority class has away node types but is not preemptible"
@@ -265,4 +315,16 @@ type WellKnownNodeType struct {
 	// Taints is the set of taints that characterizes this node type; a node is
 	// part of this node type if and only if it has all of these taints.
 	Taints []v1.Taint
+}
+
+func AllowedPriorities(priorityClasses map[string]PriorityClass) []int32 {
+	allowedPriorities := make([]int32, 0, len(priorityClasses))
+	for _, priorityClass := range priorityClasses {
+		allowedPriorities = append(allowedPriorities, priorityClass.Priority)
+		for _, awayNodeType := range priorityClass.AwayNodeTypes {
+			allowedPriorities = append(allowedPriorities, awayNodeType.Priority)
+		}
+	}
+	slices.Sort(allowedPriorities)
+	return slices.Compact(allowedPriorities)
 }
